Add files via upload
In short, the ultils_phylo script contains 2 tools. The first allows one to read JPLACE formatted files (for now only tested with output from EPA-NG) that keep tracks of the original brach numbering. The difference with the treeio and ggtree models of working with placement data is that these tools do not use the read.tree function of ape, which renumbers all the branches. Instead, a custom read function keeps track of the original branch numbers during the renumbering (necessary to comply to the phylo-object rules), and applies the new numbering to the placements as well. The second major function allows one to add multiple placements to a (phylo-object) tree that can be visualized and analyzed with common phylogenetic R-tools. As with the read function, this explicitly keeps track of the original branch numbers, as branches need to be re-numbered every time a new tip is added.